Dots can be joined, and now we use a matrix structure, rather than a vector

+++ src/controllers/input.rs
@@ -5,7 +5,7 @@ pub struct Actions {
     pub rotate_right: bool,
     pub boost: bool,
     pub shoot: bool,
-    pub click: (i32, i32),
-    pub mouse_position: (i32, i32),
+    pub click: (f64, f64),
+    pub mouse_position: (f64, f64),
     pub mouseup: bool,
 }
